fix for font crash
git-svn-id: https://dunelegacy.svn.sourceforge.net/svnroot/dunelegacy/branches/dunks@67 22675075-b818-0410-8b9e-df876f89efe1

+++ include/Font.h
@@ -46,7 +46,7 @@ class Font
         ~Font();
 
         void extents(const char* text, Uint16& w, Uint16& h);
-        void render(const char* text, SDL_Surface* surface, Uint16 x, Uint16 y, Uint8 paloff);
+        void render(const char* text, SDL_Surface* surface, int x, int y, Uint8 paloff);
 
     private:
         FNTHeader* m_header;
